Rename file (phase I).
