+++ src/commands/containers/files/openContainerFile2.ts
